Move image download fetching logic to download modal
