--- app-frontend/src/app/components/scenes/sceneDownloadModal/sceneDownloadModal.component.js
@@ -1,13 +1,0 @@
-import sceneDownloadModalTpl from './sceneDownloadModal.html';
-
-const rfSceneDownloadModal = {
-    templateUrl: sceneDownloadModalTpl,
-    bindings: {
-        close: '&',
-        dismiss: '&',
-        modalInstance: '<',
-        resolve: '<'
-    }
-};
-
-export default rfSceneDownloadModal;
